Minor fixes.
- Debugger color.
- Audio Manager null checks.
- Singleton Find.

+++ Space CUBEs Project/Assets/Little Byte/Scripts/Game Foundation/GameTime.cs
@@ -141,7 +141,10 @@ public static class GameTime
         else
         {
             timeScale = cachedTimeScale;
-            unscaledTimeJob.Kill();
+            if (unscaledTimeJob != null)
+            {
+                unscaledTimeJob.Kill();
+            }
         }
     }
